+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
-- Export as default SellerContext to use in other Apps with imports
+
+### Added
+
+- Export `CurrentSellerProvider` and `useCurrentSeller` of `CurrentSellerContext`.
 
 ## [0.1.1] - 2020-04-16
 
@@ -18,22 +20,31 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] - 2020-03-20
 
 ## [0.1.4] - 2020-04-08
+
 ### Added
+
 - README adjustments
 
 ## [0.1.3] - 2020-04-08
+
 ### Added
+
 - README adjustments
 
 ## [0.0.2] - 2020-04-08
+
 ### Added
+
 - Full documentation and Css adjustments
 
 ## [0.0.2] - 2020-03-25
+
 ### Added
+
 - Added Simple documentation
 
 ## [0.0.2] - 2020-03-20
-### Added
-- CSS Handles to every component block
 
+### Added
+
+- CSS Handles to every component block